Changes to 3hr_CMOR

+++ CMOR2/3hr_CMOR.f90
@@ -335,7 +335,7 @@ program Do3hr_CMOR
               !
               select case(ntimes(ifile,1))
               case ( 14600, 17520, 29200, 32120 )  !
-                 nchunks(ifile) = int(ntimes(ifile,1)/2920)
+                 nchunks(ifile) = ntimes(ifile,1)/2920
                  tidx1(1) =    1
                  tidx2(1) = 2920
                  do ic = 2,nchunks(ifile)
@@ -399,7 +399,7 @@ program Do3hr_CMOR
               !
               select case(ntimes(ifile,1))
               case ( 14600, 17520, 29200, 32120 )  !
-                 nchunks(ifile) = int(ntimes(ifile,1)/2920)
+                 nchunks(ifile) = ntimes(ifile,1)/2920
                  tidx1(1) =    1
                  tidx2(1) = 2920
                  do ic = 2,nchunks(ifile)
@@ -476,7 +476,7 @@ program Do3hr_CMOR
               !
               select case(ntimes(ifile,1))
               case ( 14600, 17520, 29200, 32120 )  !
-                 nchunks(ifile) = int(ntimes(ifile,1)/2920)
+                 nchunks(ifile) = ntimes(ifile,1)/2920
                  tidx1(1) =    1
                  tidx2(1) = 2920
                  do ic = 2,nchunks(ifile)
@@ -554,7 +554,7 @@ program Do3hr_CMOR
               !
               select case(ntimes(ifile,1))
               case ( 14600, 17520, 29200, 32120 )  !
-                 nchunks(ifile) = int(ntimes(ifile,1)/2920)
+                 nchunks(ifile) = ntimes(ifile,1)/2920
                  tidx1(1) =    1
                  tidx2(1) = 2920
                  do ic = 2,nchunks(ifile)
@@ -634,7 +634,7 @@ program Do3hr_CMOR
               !
               select case(ntimes(ifile,1))
               case ( 14600, 17520, 29200, 32120 )  !
-                 nchunks(ifile) = int(ntimes(ifile,1)/2920)
+                 nchunks(ifile) = ntimes(ifile,1)/2920
                  tidx1(1) =    1
                  tidx2(1) = 2920
                  do ic = 2,nchunks(ifile)
